Fix K2 Long vs Int mismatch in Baidu LocationClientOption.
setWifiCacheTimeOut takes Int; Kotlin 2.3 no longer accepts a Long from toLong().

+++ app/src/src_baidu/wangdaye/com/geometricweather/location/services/BaiduLocationService.kt
@@ -60,7 +60,7 @@ class BaiduLocationService(context: Context) : LocationService() {
         option.setIgnoreKillProcess(false)
         option.SetIgnoreCacheException(true)
         option.setEnableSimulateGps(false)
-        option.setWifiCacheTimeOut((5 * 60 * 1000).toLong())
+        option.setWifiCacheTimeOut(5 * 60 * 1000)
         val client = LocationClient(context.applicationContext)
         baiduClient = client
         client.locOption = option
